Crate slog_stream

io::Write streamer for slog-rs

Structs

AsyncStreamer
Drain formating records and writing them to a byte-stream (io::Write) asynchronously.

Streamer
Drain formating records and writing them to a byte-stream (io::Write)

Traits

Decorator
Formatted stream decorator

Format
Formats Record-s into IO

RecordDecorator
Per-record decorator

Functions

async_stream
Stream logging records to IO asynchronously

stream
Stream logging records to IO